mv constants to `constants.py`. create read_excel module. move issue sheet generation out of GUI module
