+++ packages/issue-sheet/src/d_i_gui.py
@@ -8,478 +8,20 @@ Created on Tue May 14 13:37:03 2019
 import time
 import pathlib
 import re
+import xlwings as xw
 from d_i_functions import *
+from constants import *
+from d_i_read_excel import read_excel
 #from mf_modules.tool_usage_tracking import click
 import tkinter
 from tkinter import END, RIGHT, LEFT, EXTENDED, MULTIPLE, Scrollbar, VERTICAL, Y, X, BOTH, Text, Entry, BOTTOM
-from textwrap import wrap
-from pydantic import BaseModel
+from d_i_ui import warning_messagebox, info_messagebox, getsavefilename, getfoldername, MFButton, MFLabelFrame, MFLabel, MFLabelBlack, getfilename, MFHeader, MFOptionMenu, MFCheckButton, MFTk
 
 def run():
     ''' go go go!!!!'''
     main_window = DialogWindow(None)
     main_window.mainloop()
     return True
-
-# def generate_numbers(jn):
-#     main_window = NumGeneratorWindow(None, jn)
-#     main_window.mainloop()
-#     return True
-
-class NumGeneratorWindow(MFTk):
-
-    def __init__(self, parent, jn):
-        MFTk.__init__(self, parent)
-        self.title("Document Number Generation")
-        self.sheettabledict = SHEETTABLEDICT
-        self.revisiontable = get_pandas_data() 
-        self.data = {}
-        self.job_number = jn
-        try:
-            for l in self.sheettabledict:
-                self.data[l[0]] = table_info(l[0], l[1], l[2])
-            self.listboxes = []
-            self.initialise()
-        except:
-            warning_messagebox("Document Number Generation only available in more recent versions of the spreadsheet", "Document Generator Error")
-            self._quit()
-
-    def generate(self):
-        #click("compare_docs", "DocumentIssue", self.job_number_as_int())
-        numdocs = len(self.revisiontable)
-        createddocs = 0
-        vals = []
-        for i, line in enumerate(self.sheettabledict):
-            headerrow = index_of_value(line[3], '1. Document Numbering')[0]
-            selected = self.listboxes[i].curselection()
-            if len(selected) == 0: warning_messagebox("Nothing selected in " + line[0])
-            new_val = []
-            if line[2]: #dealing with a dict
-                keys = self.data[line[0]].keys()
-                for s in selected:
-                    k = list(keys)[s]
-                    if line[0] == "drwgType":
-                        new_val.append(k)
-                    else:
-                        new_val.append(self.data[line[0]][k])
-                vals.append(new_val)
-            else:
-                index = self.data[line[0]].index
-                for s in selected:
-                    #hard code some exceptions
-                    if line[0] == "classification":
-                        new_val.append(self.data[line[0]]['classification_des'][s])
-                    else:
-                        new_val.append(index[s])
-                vals.append(new_val)
-            
-        for item in product(*vals):
-            add_doc(headerrow + createddocs + numdocs, self.sheettabledict, item)
-            createddocs += 1
-        #click("generate_numbers", "DocumentIssue", self.jn)
-
-    def initialise(self):
-        ### Banner ###
-        #MFHeader(self, text="Generate Numbers").pack(fill=BOTH)
-        MFButton(self, text="Generate Document Numbers", command=self.generate).pack(side=BOTTOM, fill=X)
-        for i, line in enumerate(reversed(self.sheettabledict)):
-            scrollbar = Scrollbar(self, orient=VERTICAL)
-            self.listboxes.insert(0, tkinter.Listbox(master=self,
-                                        height=50,
-                                        selectmode=MULTIPLE,
-                                        yscrollcommand=scrollbar.set,
-                                        exportselection=0)
-                                )
-            scrollbar.config(command=self.listboxes[0].yview)
-            scrollbar.pack(side=RIGHT, fill=Y)
-            self.listboxes[0].pack(side=RIGHT, fill=X)
-            if line[2]:
-                d = self.data[line[0]]
-                for k in d.keys():
-                    self.listboxes[0].insert(END, str(k) + " - " + str(d[k]))
-            else:
-                d = self.data[line[0]]
-                for k, item in enumerate(d.index):
-                    self.listboxes[0].insert(END, str(item) + " - " + str(d[d.columns[-1]][k]))
-
-        
-        self.protocol("WM_DELETE_WINDOW", self._quit) #Overide close event.
-
-    def _quit(self):
-        self.quit()     # stops mainloop
-        self.destroy()  # this is necessary on Windows to prevent...
-        
-        
-# get data from the spreadsheet
-class IssueSheet:
-    def __init__(self):
-        pass
-
-# mapping tables
-from typing import Any, List
-import typing as ty
-from typing_extensions import Annotated
-from pydantic.functional_validators import AfterValidator
-
-DRWG_CLASSIFICATION_CODE_REGEX = r"^[A-Z]{1}-[0-9]{2}$" # TODO: make configurable on a project basis
-UNICLASS_CLASSIFICATION_CODE_REGEX = r"^.*$" # TODO
-DrwgClassificationCode = Annotated[str, AfterValidator(lambda s: re.match(DRWG_CLASSIFICATION_CODE_REGEX, s)) is not None]
-UniclassClassificationCode = Annotated[str, AfterValidator(lambda s: re.match(DRWG_CLASSIFICATION_CODE_REGEX, s)) is not None]
-
-from typing import Annotated, Hashable, List, TypeVar
-from typing import Optional
-from pydantic_core import PydanticCustomError
-from pydantic import AfterValidator, Field, ValidationError
-from pydantic.type_adapter import TypeAdapter
-import enum
-
-T = TypeVar('T', bound=Hashable)
-
-# class DocumentCodesRegex(BaseModel):    
-#     project: str = DRWG_CLASSIFICATION_CODE_REGEX
-#     originator: str = r"^.*$"
-#     classification: str = r"^.*$"
-#     info_type: str = r"^.*$"
-#     drwg_type: str = r"^.*$"
-#     volume: str = r"^.*$"
-#     level: str = r"^.*$"
-#     sequence: str = r"^.*$"
-#
-# class DocumentCodesExtraRegex(BaseModel):
-#     classification_uniclass: str =  UNICLASS_CLASSIFICATION_CODE_REGEX
-#     classification_role: str = r"^.*$"
-#
-# NOTE: ^ maybe regex not req. as we are using enums
-
-def _validate_unique_list(v: list[T]) -> list[T]:
-    if len(v) != len(set(v)):
-        raise PydanticCustomError('unique_list', 'List must be unique')
-    return v
-
-UniqueList = Annotated[List[T], AfterValidator(_validate_unique_list), Field(json_schema_extra={'uniqueItems': True})]
-
-class DocumentCodeParts(enum.Enum):
-    project: str = "project"
-    originator: str = "originator"
-    role: str = "role"
-    classification: str = "classification"
-    type: str = "type" # info_type
-    subtype: str = "subtype" # drwg_type
-    volume: str = "volume"
-    level: str = "level"
-    sequence: str = "sequence" # remove?
-
-
-class DocumentCodesMap(BaseModel): #MapDocumentCodeDescription
-    project: dict[str, str]
-    originator: dict[ty.Literal["MXF"], ty.Literal["Max Fordham LLP"]]
-    # role: dict[str, str]
-    classification: dict[str, str]
-    type: dict[str, str]
-    # drwg_type: dict[int, str]  # info_sub_type
-    volume: dict[str, str]
-    level: dict[str, str]
-    type_sequences: Optional[dict[str, dict[str, str]]] = None  # info_type_sequences
-    
-    nested_codes: dict[DocumentCodeParts, tuple[str]] = {"classification": ("role", "subrole"), "type": ("type", "subtype")}
-    nested_codes_delimiter: str = "-"
-
-
-class DocumentMetadataMap(BaseModel):
-    scale: dict[str, str]
-    size: dict[str, str]
-    status: dict[str, str]
-    issue_format: dict[str, str]
-    doc_source: dict[str, str]
-    classification_uniclass: dict[DrwgClassificationCode, UniclassClassificationCode]
-    # role: dict[DrwgClassificationCode, str]
-    # info_sub_type: dict[str, str]
-    
-class LookupData(DocumentCodesMap, DocumentMetadataMap):
-    pass
-    
-class DocumentCodeSpec(BaseModel):
-    # regex: DocumentCodesRegex
-    order: UniqueList[DocumentCodeParts] = Field(max_length=len(DocumentCodeParts.__members__))
-    map_codes: DocumentCodesMap
-    map_metadata: DocumentMetadataMap
-    
-class Recipient(BaseModel):
-    name: str
-    email: str
-    role: str
-    
-class DistributionList(BaseModel):
-    recipients: List[Recipient]
-
-    
-class IssueSheetConfig(BaseModel):
-    pass
-    
-    
-def get_table_data(name: str) -> dict:
-    data = xw.sheets[name].range(index_of_value(f"{name}_code", name)).options(pd.DataFrame, expand='table').value
-    return data[f"{name}_des"].to_dict()
-
-def get_lookup_data() -> DocumentCodesMap: # get_lookup_data
-    names = [k for k in list(DocumentCodeParts.__members__.keys()) if k != "sequence"]
-    names = names + [stringcase.camelcase(f) for f in DocumentMetadataMap.model_fields.keys() if f not in ["classification_uniclass"]]
-    sheet_names = [s.name for s in xw.sheets]
-    names = [n for n in names if n in sheet_names]
-    data = {stringcase.snakecase(n): get_table_data(n) for n in names}
-    
-    # classification_uniclass
-    name = "classification"
-    df = xw.sheets[name].range(index_of_value(f"{name}_code", name)).options(pd.DataFrame, expand='table').value
-    data[f"classification_uniclass"] = df[f"uniclass_classification"].to_dict()
-    
-    # sequence
-    name = "sequence"
-    type_sequences = xw.sheets[name].range(index_of_value(f"{name}_code", name)).options(pd.DataFrame, expand='table').value
-    type_sequences = type_sequences.to_dict()
-    type_sequences = {k: {_k:_v for _k, _v in v.items() if _v is not None} for k, v in type_sequences.items()}
-    data["type_sequences"] = type_sequences
-    return LookupData(**data | type_sequences)
-
-import stringcase
-
-def get_issues(data):
-    '''work out which columns are dates'''
-    match_str = r"^20(\d{2}\d{2}\d{2})-.*$" # string match date format 20YYMMDD-...
-    #                                       ^ Test here: https://regex101.com/r/qH0sU7/1
-    return [c for c in data.columns if re.match(match_str, c) is not None]
-
-
-class BuildIssueSheet:
-
-    @staticmethod
-    def cols_to_plot(history=False, part=-1, max_cols_in_part=MAX_COLS_IN_PART, selected_issues=[], li_issues=[]):
-        '''get the columns to include in the pdf from the listbox'''
-
-        if part>0:
-            startindex = (part-1)*max_cols_in_part
-            endindex = part*max_cols_in_part
-            return DEFAULT_COLS + li_issues[startindex:endindex]
-
-        if history: 
-            return DEFAULT_COLS + li_issues
-
-        cols = []
-        for i in map(int, selected_issues):
-            cols.append(li_issues[i])
-        return DEFAULT_COLS + cols
-
-    @staticmethod
-    def document_number(history, part, project_code, add_space=False):  # TODO: fix this
-        ''' number has the form VWXYZ '''
-        VW = "00" #by definition
-        if history:
-            if part < 1: XYZ = "001"
-            elif part<10: XYZ = "00" + str(int(part))
-            else: XYZ = "0" + str(int(part))
-        else:
-            # XYZ = str(int(self.listbox.curselection()[0] + 100))
-            XYZ = "100"
-        doc_numb = project_code + "  -  MXF  -  XX  -  XX  -  IS  -  J  -  " + VW + XYZ
-        if not add_space: 
-            return doc_numb.replace(" ", "")
-
-        #TODO popout to allow custom document number.
-        return doc_numb
-    
-    @staticmethod
-    def construct_title_block(history, part, selected_issues, li_issues, projectinfo, config):
-        ''' title block for the document'''
-        d, m, y, s = BuildIssueSheet.day_month_year_status(-1, history, selected_issues=selected_issues, li_issues=li_issues)
-        try:
-            status_description = lookup.status[s].split(" - ")[2]
-        except:
-            status_description = ""
-            warning_messagebox(s + " - status not found in lookup table, consider checking", "Status Warning")
-
-        if history:
-            part_str = str(part)
-            if part_str:
-                doc_title = " History: Part " + part_str
-            else:
-                doc_title = " History"
-        else:
-            doc_title = "\n" + d + "/" + m + "/" + y + " - " + s
-        project_code = projectinfo.get("Project Code")
-        data = []
-        data.append([get_titleblockimage(config['office']), "", "", "client", "", "", "project", "", "", "document title", "", ""])
-        data.append(["", "", "", projectinfo.get("Client Name"), "", "", projectinfo.get("Project Name"), "", "", "Document Issue Sheet" + doc_title, "", ""])
-        data.append(["", "", "", "job no.", "project leader", "scale at A3", "", "", "", "", "", ""])
-        data.append(["", "", "",  projectinfo.get("Job Number"), projectinfo.get("Project Leader") ,"NTS", "", "", "", "", "", ""])
-        data.append(["", "", "", "status code and description", "", "", "issue date", "classification", "revision", projectinfo.get("Naming Convention"), "", ""])
-        data.append(["", "", "", s + status_description, "", "", d + "/" + m + "/" + y, "-", "-", BuildIssueSheet.document_number(history, part, project_code, add_space=True), "", ""])
-        return data    
-
-
-    @staticmethod
-    def day_month_year_status(index, history, selected_issues=[], li_issues=[]):
-        cols_to_plot = BuildIssueSheet.cols_to_plot(history=history,
-                 selected_issues=selected_issues,
-                 li_issues=li_issues)[index]
-        d = cols_to_plot[6:8]
-        m = cols_to_plot[4:6]
-        y = cols_to_plot[2:4]
-        s = cols_to_plot[9:]
-        return d, m, y, s
-
-    @staticmethod
-    def data_table(history, part, selected_issues, li_issues, data):       
-        cols = BuildIssueSheet.cols_to_plot(history=history,
-                 selected_issues=selected_issues, 
-                 li_issues=li_issues, part=part)
-        last_col = cols[-1]
-        data_tmp = data.sort_values("System Identifier Description")
-        data_list = [] #this is a list of rows in the table. #list for styling output.
-        sid_style = []
-
-        uniclass_classifications = sorted(list(set([d["uniclass"] for d in doc_descriptions.values()])))
-        map_uniclass_description =  {v: lookup.classification[k] for k, v in lookup.classification_uniclass.items()}
-            
-        for uniclass in uniclass_classifications:
-            sid = map_uniclass_description[uniclass]
-            if uniclass == "N/A": 
-                uniclass = ""
-
-            df = data_tmp[data_tmp["System Identifier Description"] == sid].sort_values('Document Number')
-            if history: 
-                mask = df["Current Rev"].str.len() > 0
-            else: 
-                mask = df[last_col].str.len() >= 1
-
-            if df.loc[mask, cols].values.tolist():
-                data_list += [[p_nospace("{0}    {1}".format(sid,uniclass), [])]]
-                sid_style += sid_line_style(4+len(data_list))
-                data_list += format_data_rows(df.loc[mask, cols].values.tolist()) #This is where they get added.
-                data_list += [[""]*len(cols)]
-
-        return data_list, sid_style
-
-    @staticmethod
-    def tablestyle(history, sid_style, data_list, headings):
-        if history:
-            tablestyle = highlight_last_format(headings+data_list, startrow=5, rev_position=len(DEFAULT_COLS)-1) + sid_style
-        else:
-            tablestyle = DEFAULTTABLESTYLE(defaultcols = len(DEFAULT_COLS)-1) + sid_style
-        return tablestyle
-
-    @staticmethod
-    def distribution_table(dist_data, data_list, sid_style, history, part, selected_issues, li_issues):
-        ###LET's Do the Distribution List
-        cols = BuildIssueSheet.cols_to_plot(history=history,
-                 selected_issues=selected_issues, 
-                 li_issues=li_issues, part=part)
-        
-        data_list += [[""]] #blank line
-        sid_style += dist_line_style(4+len(data_list))
-        data_list += [["Distribution"]]
-        sid_style += dist_line_style(4+len(data_list))
-        distcols = ['Name']*len(DEFAULT_COLS) + [x for x in cols if not x in DEFAULT_COLS]
-        for i, line in enumerate(dist_data[distcols].values.tolist()):
-            sid_style.append(('SPAN', (0, 5+len(data_list)+i), (len(DEFAULT_COLS)-1, 5+len(data_list)+i)))
-        data_list += dist_data[distcols].values.tolist()
-        return data_list, sid_style
-
-    @staticmethod
-    def create_heading(cols_to_plot):
-        ''' format the thing that goes at the top '''
-        headings = [[""]*len(cols_to_plot) + [""],
-                    [""]*len(cols_to_plot),
-                    [""]*len(cols_to_plot),
-                    [""]*len(cols_to_plot),
-                    cols_to_plot]
-
-        for i in range(len(DEFAULT_COLS), len(cols_to_plot)):
-            headings[0][i] = cols_to_plot[i][6:8]
-            headings[1][i] = cols_to_plot[i][4:6]
-            headings[2][i] = cols_to_plot[i][2:4]
-            headings[3][i] = cols_to_plot[i][9:]
-            headings[4][i] = ""
-
-        for i, string in enumerate(["Day", "Month", "Year", "Status"]):
-            headings[i][len(DEFAULT_COLS)-1] = string
-
-        for i in range(1, len(DEFAULT_TITLES)):
-            headings[4][i] = DEFAULT_TITLES[i]
-        return headings
-    
-    
-    @staticmethod
-    def output_doc(document, fname, data, projectinfo, config, history=False, part=-1, selected_issues=[], column_widths=[100, 40, 9], li_issues=[]):
-
-        document.filename = fname
-        if not document.filename: 
-            return False
-        #data_list is a list of lists containing the cells of the final table.
-        #sid_style is the additional styling applied to the cells over and above the standard.
-        
-        data_list, sid_style = BuildIssueSheet.data_table(history, part, selected_issues, li_issues, data=data)
-        cols_to_plot = BuildIssueSheet.cols_to_plot(history=history,
-                 selected_issues=selected_issues, 
-                 li_issues=li_issues, part=part)
-        headings = BuildIssueSheet.create_heading(cols_to_plot)
-        data_list, sid_style = BuildIssueSheet.distribution_table(dist_data, data_list, sid_style, history, part, selected_issues, li_issues)
-        titleblockdata = BuildIssueSheet.construct_title_block(history, part, selected_issues, li_issues, projectinfo, config)
-        tstyle = BuildIssueSheet.tablestyle(history, sid_style, data_list, headings)
-        document = issue_sheet(data_list,
-                                document,
-                                titleblockdata = titleblockdata,
-                                headings=headings,
-                                tablestyle=tstyle,
-                                col_widths = column_widths[:3])  # This saves the file to pdf
-        return True
-    
-    @staticmethod
-    def print_issue_and_issue_history(data, projectinfo, config, selected_issues, column_widths, project_name, office, max_cols_in_part, li_issues, project_code) -> tuple[str, str]:
-        filescreated = []
-        document = new_document(project_name, office)
-        no_issues = len(li_issues)
-        #output doc.
-        savefilename = getsavefilename(extension="pdf", initialfile="IssueSheet.pdf")
-        savefilenamehistory = savefilename.replace(".pdf", "_history " + BuildIssueSheet.document_number(True, -1, project_code) + ".pdf")
-        fname = savefilename.replace(".pdf", " " + BuildIssueSheet.document_number(False, -1, project_code) + ".pdf")
-        
-        # output issue sheet
-        if not BuildIssueSheet.output_doc(document, fname, data, projectinfo, config, selected_issues=selected_issues, li_issues=li_issues, column_widths=column_widths): 
-            return False
-        else:
-            filescreated.append(savefilename.replace(".pdf", " " + BuildIssueSheet.document_number(False, -1, project_code) + ".pdf"))
- 
-        # continue to output issue istory
-        #output history
-        if no_issues > max_cols_in_part: #Pagination required.
-            num_parts = math.ceil(no_issues/max_cols_in_part)
-            for part in range(1, num_parts+1):
-                fname = savefilename.replace(".pdf", "_history " + BuildIssueSheet.document_number(True, part, project_code) + ".pdf")
-                if not BuildIssueSheet.output_doc(document, fname, data, projectinfo, config, history=True, part=part, selected_issues=selected_issues, li_issues=li_issues, column_widths=column_widths): 
-                    return False
-                else: 
-                    filescreated.append(fname)
-            savefilenamehistory = filescreated[-1]
-        else:
-            if not BuildIssueSheet.output_doc(document, savefilenamehistory, data, projectinfo, config, history=True, selected_issues=selected_issues, li_issues=li_issues, column_widths=column_widths): 
-                return False
-            else: 
-                filescreated.append(savefilenamehistory)
-                
-        return filescreated
-
-
-
-lookup = get_lookup_data()
-projectinfo = project_info()
-config = user_config(projectinfo.get("Job Number")) # define this in the UI
-data = get_pandas_data()
-li_issues = get_issues(data)
-doc_revs = data["Current Rev"].to_dict()
-doc_descriptions = data[[c for c in data.columns if c not in li_issues+["Current Rev"]]].T.to_dict()
-doc_issues = data[li_issues].T.to_dict()
-dist_data = get_distribution_data(li_issues=li_issues)
-print("data loaded")
 
 
 ### THE INTERFACE fpor the wizard###
@@ -498,6 +40,10 @@ class DialogWindow(MFTk):
         self.headers = self.data.columns
         self.dist_data = get_distribution_data()
         self.fix_dist_data()
+        
+        
+        self.lookup, self.projectinfo, self.config, self.data, self.li_issues, self.doc_revs, self.doc_descriptions, self.doc_issues, self.dist_data = read_excel()
+        
         
         self.col_widths = tkinter.StringVar(self)
         try: #try for backwards compat.
@@ -525,13 +71,6 @@ class DialogWindow(MFTk):
         self.last_col = None
         self.initialise()
         self.protocol("WM_DELETE_WINDOW", self._quit) #Ovveride close event.
-
-    def job_number_as_int(self):
-        jn = self.get_project_info("number")
-        try:
-            return int(jn.lower().replace("j", ""))
-        except:
-            return 4321
 
     def get_project_info(self, key):
         try:
@@ -677,7 +216,6 @@ class DialogWindow(MFTk):
 
     def compare_docs(self):
         ''' check that all the files that are in the issue sheet are in the outgoing folder'''
-        #click("compare_docs", "DocumentIssue", self.job_number_as_int())
         if not self.outgoingfolder:
             info_messagebox(message="No Outgoing folder selected")
             return False
@@ -697,7 +235,7 @@ class DialogWindow(MFTk):
 
         cols = BuildIssueSheet.cols_to_plot(history=False,
                  selected_issues=self.listbox.curselection(), 
-                 li_issues=li_issues)
+                 li_issues=self.li_issues)
         self.last_col = last_col = cols[-1]
         if last_col not in self.dates(): warning_messagebox("No Specific Issue selected", "Input error")
         mask = self.data[last_col].str.len() >= 1
@@ -743,8 +281,9 @@ class DialogWindow(MFTk):
             if not self.compare_docs(): 
                 return False
             
-        
-        filescreated = BuildIssueSheet.print_issue_and_issue_history(self.data, self.projectinfo, self.config, selected_issues, column_widths, project_name, office, max_cols_in_part, li_issues, project_code)
+        savefilename = getsavefilename(extension="pdf", initialfile="IssueSheet.pdf")
+        savefilenamehistory = savefilename.replace(".pdf", "_history " + BuildIssueSheet.document_number(True, -1, project_code) + ".pdf")
+        filescreated = BuildIssueSheet.print_issue_and_issue_history(self.data, self.dist_data, self.projectinfo, self.lookup, self.config, self.doc_descriptions, selected_issues, column_widths, project_name, office, max_cols_in_part, self.li_issues, project_code)
         self.savefilename, self.savefilenamehistory = filescreated[0], filescreated[1]
 
         info_messagebox(message="Documents saved:\n" + "\n".join(filescreated))
